// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Study = {
  ELSYS: 'ELSYS',
  KYB: 'KYB',
  OTHER: 'OTHER',
  NOTANSWERD: 'NOTANSWERD'
} as const

export type Study = (typeof Study)[keyof typeof Study]


export const AuksjonsObjektType = {
  AUKSJON: 'AUKSJON',
  SALG: 'SALG',
  LIVE: 'LIVE'
} as const

export type AuksjonsObjektType = (typeof AuksjonsObjektType)[keyof typeof AuksjonsObjektType]


export const SpecialObjectType = {
  OVERRIDE: 'OVERRIDE'
} as const

export type SpecialObjectType = (typeof SpecialObjectType)[keyof typeof SpecialObjectType]


export const Committee = {
  NOTCOM: 'NOTCOM',
  VEVCOM: 'VEVCOM',
  HS: 'HS',
  FADDERSTYRET: 'FADDERSTYRET',
  OMBUL: 'OMBUL',
  LOCCOM: 'LOCCOM',
  CONTACTOR: 'CONTACTOR',
  KIELDER: 'KIELDER',
  PHAESTCOM: 'PHAESTCOM',
  SPORTOGSPILL: 'SPORTOGSPILL',
  DGR: 'DGR',
  OV: 'OV',
  SOSCOM: 'SOSCOM',
  HEUTTECOM: 'HEUTTECOM',
  BRYGCOM: 'BRYGCOM',
  BLAESTCOM: 'BLAESTCOM',
  LOPHTCOM: 'LOPHTCOM',
  OMEGAREVYEN: 'OMEGAREVYEN',
  PHINANSCOM: 'PHINANSCOM'
} as const

export type Committee = (typeof Committee)[keyof typeof Committee]